// Features/reports/Types/INetPayForReport.ts
export interface INetPayForReport {
    identifier: string;
    idEmployee: number;
    employeeName: string;
    salary: number;
    idConcept: number;
    conceptCode: string;
    concept: string;
    amount: number;
    idPayrollPay: number;
    payrollName: string;
    isProfit: boolean;
    payrollPayDate: string;
    idCompany: number;
}